requirements.txt
fastapi
uvicorn
python-telegram-bot==21.6
psycopg[binary]
pydantic
supabase
python-dateutil
